./ bug and hardlink folder

+++ src/mx_print_total.c
@@ -7,6 +7,6 @@ const char *mx_print_total(t_elem **arr, t_all *ptr) {
     for (int k = 0; k < ptr->count; k++) {
         total += arr[k]->b;
     }
-    res = mx_itoa((int)total);
+    res = mx_itoa(total);
     return res;
 }
